fixed a bug in libcrispr.cpp caused by bad casting

+++ src/crass/libcrispr.cpp
@@ -915,7 +915,6 @@ bool qcFoundRepeats(ReadHolder * tmp_holder, int minSpacerLength, int maxSpacerL
         std::vector<std::string>::iterator spacer_last = spacer_vec.end();
         spacer_iter += sp_start_offset;
         spacer_last -= sp_end_offset;
-        
         while (spacer_iter != spacer_last) 
         {
         	num_compared++;
@@ -923,9 +922,8 @@ bool qcFoundRepeats(ReadHolder * tmp_holder, int minSpacerLength, int maxSpacerL
             float ss_diff = PatternMatcher::getStringSimilarity(*spacer_iter, *(spacer_iter + 1));
             ave_spacer_to_spacer_difference += ss_diff;
             //MI std::cout << ss_diff << " : " << *spacer_iter << " : " << *(spacer_iter + 1) << std::endl;
-            
-            ave_spacer_to_spacer_len_difference += spacer_iter->size() - (spacer_iter + 1)->size();
-            ave_repeat_to_spacer_len_difference += repeat.size() - spacer_iter->size();
+            ave_spacer_to_spacer_len_difference += ((float)spacer_iter->size() - (float)(spacer_iter + 1)->size());
+            ave_repeat_to_spacer_len_difference +=  ((float)repeat.size() - (float)spacer_iter->size());
             
             if((int)(spacer_iter->length()) < min_spacer_length)
             {
